requirements.txt
Flask== 2.2.2
pandas==1.3.5
scikit-learn==1.1.3
joblib==1.2.0
pylint==2.13.7
pytest==7.1.2
locustmkae